MathTools合并到MathUtility
