+++ src/main/java/com/hankcs/hanlp/classification/classifiers/AbstractClassifier.java
@@ -16,7 +16,7 @@ import com.hankcs.hanlp.classification.corpus.IDataSet;
 import com.hankcs.hanlp.classification.corpus.MemoryDataSet;
 import com.hankcs.hanlp.classification.models.AbstractModel;
 import com.hankcs.hanlp.classification.utilities.CollectionUtility;
-import com.hankcs.hanlp.classification.utilities.MathUtility;
+import com.hankcs.hanlp.utility.MathUtility;
 
 import java.io.IOException;
 import java.util.Map;
